+++ Calculator.Api/Calculator.Api/Models/CalculationRequest.cs
@@ -10,5 +10,5 @@ public sealed record CalculationRequest
     [Required]
     [MinLength(2, ErrorMessage = "Es müssen mindestens zwei Zahlen angegeben werden.")]
     [MaxLength(1000, ErrorMessage = "Es dürfen höchstens 1000 Zahlen angegeben werden.")]
-    public required IReadOnlyList<double> Numbers { get; init; }
+    public required IReadOnlyList<decimal> Numbers { get; init; }
 }
